Change delete confirmation to use modal instead of in sidedrawer

+++ src/components/DataCrudFormPage/DataCrudFormPage.tsx
@@ -15,6 +15,9 @@ import {FormActionType} from '../../state/crud-page/form-action-type';
 import {GridRowData} from '../Grid/grid-row-data';
 import DeleteCellRenderer from '../DeleteCellRenderer/DeleteCellRenderer';
 import GridColumn from '../Grid/GridColumn';
+import Modal from '../Modal/Modal';
+import ModalTitle from '../Modal/ModalTitle';
+import ModalFooterSubmit from '../Modal/ModalFooterSubmit';
 
 /***
  * Generic page template for CRUD operations on entity arrays.
@@ -26,18 +29,6 @@ export function DataCrudFormPage<T extends GridRowData, R> (props: DataCrudFormP
   const dataState: DataService<any, any> = props.useDataState();
 
   const pageState: State<CrudPageState<any>> = props.usePageState();
-
-  const DeleteComponent = props.deleteComponent;
-  const deleteBtnName = 'Delete';
-  let columns: GridColumn[];
-  if (props.allowDelete && DeleteComponent) {
-    columns = [
-      ...props.columns,
-      new GridColumn('', false, false, deleteBtnName, 'header-center', DeleteCellRenderer, { onClick: deleteConfirmation })
-    ];
-  } else {
-    columns = props.columns;
-  }
 
   useEffect(() => {
     dataState.fetchAndStoreData();
@@ -52,20 +43,20 @@ export function DataCrudFormPage<T extends GridRowData, R> (props: DataCrudFormP
       const rowData = event.data;
       if (rowData != null) {
         const dtoData = await dataState.convertRowDataToEditableData(rowData);
-        pageState.set({
+        pageState.merge({
           formAction: FormActionType.UPDATE,
           isOpen: true,
           selected: dtoData,
           formErrors: undefined,
           successAction: undefined,
-          isSubmitting: false,
+          isSubmitting: false
         });
       }
     }
   }
 
   function onAddEntityClick() {
-    pageState.set({
+    pageState.merge({
       formAction: FormActionType.ADD,
       isOpen: true,
       selected: undefined,
@@ -82,13 +73,13 @@ export function DataCrudFormPage<T extends GridRowData, R> (props: DataCrudFormP
   async function deleteConfirmation(deleteItem: R) {
     if (props.allowDelete && deleteItem != null) {
       const data = await dataState.convertRowDataToEditableData(deleteItem);
-      pageState.set({
-        formAction: FormActionType.DELETE,
-        isOpen: true,
+
+      pageState.merge({
+        isDeleteConfirmationOpen: true,
         selected: data,
         formErrors: undefined,
         successAction: undefined,
-        isSubmitting: false,
+        isSubmitting: false
       });
     }
   }
@@ -96,34 +87,30 @@ export function DataCrudFormPage<T extends GridRowData, R> (props: DataCrudFormP
   async function deleteSubmit() {
     const deleteItem = pageState.selected.get();
 
-    pageState.set(prevState => ({
-      ...prevState,
-      isSubmitting: true
-    }));
+    if (deleteItem == null)
+      return;
+
+    pageState.isSubmitting.set(true);
 
     try {
       await dataState.sendDelete(deleteItem);
 
-      pageState.set(prevState => {
-        return {
-          ...prevState,
-          successAction: {
-            success: true,
-            successMsg: `Successfully deleted ${props.dataTypeName}.`,
-          },
-          isSubmitting: false
-        }
+      pageState.merge({
+        isDeleteConfirmationOpen: false,
+        successAction: {
+          success: true,
+          successMsg: `Successfully deleted ${props.dataTypeName}.`
+        },
+        isSubmitting: false
       });
+
     }
     catch (error) {
-      pageState.set(prevState => {
-        return {
-          ...prevState,
-          formErrors: {
-            general: error.message
-          },
-          isSubmitting: false
-        }
+      pageState.merge({
+        formErrors: {
+          general: error.message
+        },
+        isSubmitting: false
       });
     }
   }
@@ -194,8 +181,21 @@ export function DataCrudFormPage<T extends GridRowData, R> (props: DataCrudFormP
 
   const UpdateForm = props.updateForm;
   const CreateForm = props.createForm;
+  const DeleteComponent = props.deleteComponent;
+
+  const deleteBtnName = 'Delete';
+  let columns: GridColumn[];
+  if (props.allowDelete && DeleteComponent) {
+    columns = [
+      ...props.columns,
+      new GridColumn('', false, false, deleteBtnName, 'header-center', DeleteCellRenderer, { onClick: deleteConfirmation })
+    ];
+  } else {
+    columns = props.columns;
+  }
 
   return (
+    <>
       <PageFormat pageTitle={props.pageTitle}>
         <Container fluid style={{ height: '100%' }}>
           {dataState.isPromised ?
@@ -225,45 +225,55 @@ export function DataCrudFormPage<T extends GridRowData, R> (props: DataCrudFormP
                       />
 
                       <SideDrawer title={props.dataTypeName} isOpen={pageState.isOpen.get()} onCloseHandler={onCloseHandler}>
-                  {
-                    pageState.formAction.value === FormActionType.ADD ?
-                      <CreateForm
-                        onSubmit={createSubmit}
-                        formActionType={FormActionType.ADD}
-                        formErrors={pageState.formErrors.get()}
-                        onClose={onCloseHandler}
-                        successAction={pageState.successAction.get()}
-                        isSubmitting={pageState.isSubmitting.get()}
-                      />
-                      : pageState.formAction.value === FormActionType.UPDATE && pageState.selected.get() ?
-                        <UpdateForm
-                          data={pageState.selected.get()}
-                          formErrors={pageState.formErrors.get()}
-                          onSubmit={updateSubmit}
-                          onClose={onCloseHandler}
-                          successAction={pageState.successAction.get()}
-                          isSubmitting={pageState.isSubmitting.get()}
-                          formActionType={FormActionType.UPDATE}
-                        />
-                        : pageState.selected.get() && pageState.formAction.value === FormActionType.DELETE && DeleteComponent ?
-                          <DeleteComponent
-                            data={pageState.selected.get()}
+                        {
+                          pageState.formAction.value === FormActionType.ADD ?
+                          <CreateForm
+                            onSubmit={createSubmit}
+                            formActionType={FormActionType.ADD}
                             formErrors={pageState.formErrors.get()}
-                            onSubmit={deleteSubmit}
                             onClose={onCloseHandler}
                             successAction={pageState.successAction.get()}
                             isSubmitting={pageState.isSubmitting.get()}
-                          /> 
-                          : null
+                          />
+                          : pageState.formAction.value === FormActionType.UPDATE && pageState.selected.get() ?
+                            <UpdateForm
+                              data={pageState.selected.get()}
+                              formErrors={pageState.formErrors.get()}
+                              onSubmit={updateSubmit}
+                              onClose={onCloseHandler}
+                              successAction={pageState.successAction.get()}
+                              isSubmitting={pageState.isSubmitting.get()}
+                              formActionType={FormActionType.UPDATE}
+                            />
+                            : null
                         }
                       </SideDrawer>
                     </>
                 }
-
               </div>
           }
         </Container>
       </PageFormat>
 
+      { props.allowDelete && DeleteComponent && pageState.selected.get() &&
+        <Modal
+          headerComponent={<ModalTitle title="Delete Confirmation" />}
+          footerComponent={<ModalFooterSubmit 
+                              onCancel={onCloseHandler} 
+                              onSubmit={deleteSubmit} 
+                              disableSubmit={pageState.isSubmitting.get() || pageState.successAction.get()?.success} 
+                            />}
+          show={pageState.isDeleteConfirmationOpen.get()}
+          onHide={onCloseHandler}
+        >
+          <DeleteComponent
+            data={pageState.selected.get()}
+            formErrors={pageState.formErrors.get()}
+            successAction={pageState.successAction.get()}
+            isSubmitting={pageState.isSubmitting.get()}
+          />
+        </Modal>
+      }
+    </>
   )
 }
